+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## TODO
+-
+
+## Release 2022-11-11
+- Fixed a bug in ElectionDay calendar, was incorrectly calculating a date on 1-Nov 
+which is not allowed
+- Set up the Ercot ISO.  Added Ercot specific buckets.
+
 ## Release 2022-10-31
 - Add a name to the holiday type enum
 - Bump SDK requirement to 2.17